+++ test/fixtures/cookbooks/pgtest/recipes/install_ext.rb
@@ -1,15 +1,16 @@
 include_recipe 'postgresql_lwrp::default'
 
 postgresql_extension 'cube' do
-  in_version node['postgresql']['defaults']['server']['version']
+  in_version node['pgtest']['version']
   in_cluster 'main'
   db 'test01'
 end
 
-pgxn_extension 'count_distinct' do
-  in_version node['postgresql']['defaults']['server']['version']
+pgxn_extension 'semver' do
+  in_version node['pgtest']['version']
   in_cluster 'main'
   db 'test01'
-  version '1.3.2'
+  version '0.20.3'
   stage 'stable'
+  only_if { node['pgtest']['version'].to_f > 9.1 }
 end