+++ test/integration/postgresql_lwrp_test/inspec/libraries/postgres_extension.rb
@@ -5,18 +5,15 @@ class PostgresExtension < Inspec.resource(1)
   name 'postgres_extension'
   desc 'Use the postgres_extension InSpec audit resource to test installation of PostgreSQL database extensions'
   example "
-    describe postgres_extension('cube', ['test01']) do
+    describe postgres_extension('9.6', 'main', 'cube', 'test01') do
       it { should be_installed }
     end
   "
 
-  def initialize(name, db = ['postgres'], host = 'localhost', port = nil)
+  def initialize(version, cluster, name, db = 'postgres')
     @name = name
     @db = db
-    version = version_from_psql
-    cluster = cluster_from_dir("/etc/postgresql/#{version}")
-    @host = host
-    @port = port || get_port(version, cluster)
+    @port = get_port(version, cluster)
   end
 
   def installed?
@@ -29,11 +26,6 @@ class PostgresExtension < Inspec.resource(1)
   end
 
   private
-
-  def get_port(version, cluster)
-    postmaster_content = inspec.command("cat /var/lib/postgresql/#{version}/#{cluster}/postmaster.pid").stdout.split
-    postmaster_content[3].to_i
-  end
 
   def query(query)
     psql_cmd = create_psql_cmd(query, @db)
@@ -50,27 +42,13 @@ class PostgresExtension < Inspec.resource(1)
     Shellwords.escape(query)
   end
 
-  def create_psql_cmd(query, db = [])
-    dbs = db.map { |x| "-d #{x}" }.join(' ')
-    "su postgres -c \"psql #{dbs} -p #{@port} -q  -t -c #{escaped_query(query)}\""
+  # TODO: You cannot specify multiple DBs
+  def create_psql_cmd(query, db)
+    "su postgres -c \"psql -d #{db} -p #{@port} -q -t -c #{escaped_query(query)}\""
   end
 
-  def version_from_psql
-    return unless inspec.command('psql').exist?
-    inspec.command("psql --version | head -n 1 | awk '{ print $NF }' | awk -F. '{ print $1\".\"$2 }'").stdout.strip
-  end
-
-  def cluster_from_dir(dir)
-    if inspec.directory("#{dir}/main").exist?
-      'main'
-    else
-      dirs = inspec.command("ls -d #{dir}/*/").stdout.lines
-      first = dirs.first.chomp.split('/').last
-      if dirs.count > 1
-        warn "Multiple postgresql clusters configured or incorrect base dir #{dir}"
-        warn "Using the first directory found: #{first}"
-      end
-      first
-    end
+  def get_port(version, cluster)
+    postmaster_content = inspec.command("cat /var/lib/postgresql/#{version}/#{cluster}/postmaster.pid").stdout.split
+    postmaster_content[3].to_i
   end
 end